Update onContentEdit.cfm
Problem: When editing content and attempting to assign/create a translation mapping the user can't get past the "Save as Draft" dialog.  The dialog functions properly, but the user is never redirected to the plugin for the purpose of creating a translation mapping. Instead the edit screen they were on merely refreshes, taking them essentially no where. 

This change allows the user to actually enter the plugin after making their choice in the "Save as draft" dialog.